note_checker/pages/requirements.txt
streamlit
soundfile
requests
Pillow
pytesseract
docx
opencv-python
numpy
pdfplumber
moviepy
pydub
noisereduce
SpeechRecognition